refactor(usage): keep cache state private and test its owner
