+++ src/gateway/server-methods/usage.sessions-usage-owner-attribution.test.ts
@@ -34,7 +34,7 @@ import {
   loadSessionCostSummariesFromCache,
 } from "../../infra/session-cost-usage.js";
 import { loadCombinedSessionStoreForGatewayCore } from "../session-utils.js";
-import { testApi, usageHandlers } from "./usage.js";
+import { usageHandlers } from "./usage.js";
 
 type StoredFixture = { key: string; agentId: string; entry: SessionEntry };
 type DiscoveredFixture = { agentId: string; sessionId: string };
@@ -67,6 +67,7 @@ async function queryUsage(options: {
   config?: OpenClawConfig;
   params?: Record<string, unknown>;
 }): Promise<SessionsUsageResult> {
+  const runtimeConfig = { ...(options.config ?? defaultConfig) };
   return await withTempDir("usage-owner-", async (stateDir) =>
     withEnvAsync({ OPENCLAW_STATE_DIR: stateDir }, async () => {
       const store = Object.fromEntries(options.rows.map(({ key, entry }) => [key, entry]));
@@ -120,7 +121,7 @@ async function queryUsage(options: {
         )({
           respond,
           params: { range: "all", limit: 10, agentScope: "all", ...options.params },
-          context: { getRuntimeConfig: () => options.config ?? defaultConfig },
+          context: { getRuntimeConfig: () => runtimeConfig },
         } as unknown as Parameters<(typeof usageHandlers)["sessions.usage"]>[0]);
         expect(respond).toHaveBeenCalledTimes(1);
         const call = expectDefined(respond.mock.calls[0], "sessions.usage response");
@@ -158,7 +159,6 @@ function expectRows(
 
 describe("sessions.usage owner attribution", () => {
   beforeEach(() => {
-    testApi.sessionsUsageCache.clear();
     vi.clearAllMocks();
   });
