+++ test/e2e/qa-lab/runtime/gateway-usage-memory-apis.e2e.test.ts
@@ -13,7 +13,6 @@ import {
 import type { OpenClawConfig } from "../../../../src/config/types.openclaw.js";
 import { READ_SCOPE } from "../../../../src/gateway/method-scopes.js";
 import { clearModelAuthStatusUsageCache } from "../../../../src/gateway/server-methods/models-auth-status-usage-cache.js";
-import { testApi as usageTestApi } from "../../../../src/gateway/server-methods/usage.js";
 import { startGatewayServer } from "../../../../src/gateway/server.js";
 import { loadGatewaySessionEntryReadOnly } from "../../../../src/gateway/session-utils.js";
 import {
@@ -191,8 +190,6 @@ describe("gateway usage and memory APIs", () => {
         clearRuntimeConfigSnapshot();
         clearConfigCache();
         clearModelAuthStatusUsageCache();
-        usageTestApi.costUsageCache.clear();
-        usageTestApi.sessionsUsageCache.clear();
 
         const { databasePath } = await seedCompletedUsageSession(state);
         const databaseStats = await fs.stat(databasePath);
@@ -336,8 +333,6 @@ describe("gateway usage and memory APIs", () => {
         }
         await server?.close({ reason: "gateway usage and memory QA complete" });
         clearModelAuthStatusUsageCache();
-        usageTestApi.costUsageCache.clear();
-        usageTestApi.sessionsUsageCache.clear();
         await state.cleanup();
       }
     },
